feat(pull-event.usecase): implement the pulling event use case
